Remember last session if not committed
In case the user closes without committing, this is a hint that she
forgot something and wants to commit after some more changes. Thus we
keep the title and message kin that case so no work is lost.

Only after a successful commit we clear the title and message on the
next program start because typing the exact same message again is very
uncommon.