Show given message if any; guess comment character
If committing with a pre-defined message, the editor starts with that
message. For example, the command

    git commit -m "This is the first line" -e

will open the editor, because of the -e option, and have a message
already. This message is now used to initialize the editor.

The comment character used to be assumed to be '#' but now a heuristic
reads it from the message.